pyproject.toml
[tool.poetry]
name = "llm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.2"
pillow = "^12.1.1"
black = "^26.3.1"
djangorestframework-simplejwt = "^5.5.1"
coverage = "^7.13.5"
drf-yasg = "^1.21.15"
stripe = "^15.0.0"
redis = "^7.4.0"
eventlet = "^0.40.4"
celery = "^5.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
